# api/openapi.yaml
openapi: 3.0.3
info:
  title: Sport Track API
  version: 1.0.0
  description: |
    API REST de Sport Track (suivi sportif coach / sportif).

    **Authentification** : `POST /auth/login` (ou `/auth/google`) renvoie un jeton
    à envoyer dans `Authorization: Bearer <jeton>`.

    **Rôles** : `USER` (sportif), `COACH`, `ADMIN`. COACH et ADMIN sont des coachs
    (« staff ») : ils créent les séances, notent les sportifs et ont des abonnés.

    **Erreurs** : `{ "error": "message lisible", "code": "CODE_METIER" }`. Le `code`
    est stable (ex. `UNAUTHENTICATED`, `FORBIDDEN`, `VALIDATION_ERROR`,
    `ATHLETE_NOT_SUBSCRIBED`, `TEMPLATE_NOT_FOUND`).
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /profils:
    get:
      tags: [Profils]
      summary: Liste de tous les comptes (admin uniquement)
      responses:
        "200":
          description: Profils
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Profil" } }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [Profils]
      summary: Inscription, ou création d'un sportif par un coach
      description: Public. Un sportif créé par un coach connecté lui est abonné d'office. Le rôle ADMIN est réservé à un admin.
      security: [{}, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfilInput" }
      responses:
        "201":
          description: Profil créé
          content: { application/json: { schema: { $ref: "#/components/schemas/Profil" } } }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
  /circuits:
    get:
      tags: [Séances]
      summary: Séances visibles par l'utilisateur
      description: Un sportif voit ses séances, un coach les siennes, un admin toutes. Les filtres ne peuvent que restreindre ce périmètre.
      parameters:
        - { name: profilId, in: query, schema: { type: integer } }
        - { name: auteurId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Séances
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Circuit" } }
    post:
      tags: [Séances]
      summary: Créer une séance (coach)
      description: |
        `niveau` est un texte libre (valeur prédéfinie ou difficulté personnalisée).
        Le nom d'un exercice est libre (exercice du catalogue ou saisi par le coach).
        Chaque participant doit être un abonné actif du coach (403 `ATHLETE_NOT_SUBSCRIBED` sinon).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/CircuitInput"
                - type: object
                  properties:
                    participantIds: { type: array, items: { type: integer } }
                    exercices: { type: array, items: { $ref: "#/components/schemas/ExerciceInput" } }
      responses:
        "201":
          description: Séance créée
          content: { application/json: { schema: { $ref: "#/components/schemas/Circuit" } } }
        "403": { $ref: "#/components/responses/Error" }
...
  /exercices:
    get:
      tags: [Exercices]
      summary: Exercices d'une séance gérée par le coach
      parameters:
        - { name: circuitId, in: query, description: Obligatoire sauf pour un admin, schema: { type: integer } }
      responses:
        "200":
          description: Exercices
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Exercice" } }
    post:
      tags: [Exercices]
      summary: Ajouter un exercice à une séance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ExerciceInput"
                - type: object
                  required: [circuitId]
                  properties:
                    circuitId: { type: integer }
      responses:
        "201":
          description: Exercice créé
          content: { application/json: { schema: { $ref: "#/components/schemas/Exercice" } } }
...
  /templates:
    get:
      tags: [Templates]
      summary: Templates du coach connecté
      responses:
        "200":
          description: Résumé des templates, du plus récemment modifié au plus ancien
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/TemplateSummary" } }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [Templates]
      summary: Créer un template
      description: Le propriétaire est le coach du jeton. Aucun participant n'est enregistré.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TemplateInput" }
      responses:
        "201":
          description: Template créé
          content: { application/json: { schema: { $ref: "#/components/schemas/Template" } } }
        "400": { $ref: "#/components/responses/Error" }
...
  /coaches:
    get:
      tags: [Abonnements]
      summary: Rechercher des coachs
      parameters:
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Coachs, avec l'état d'abonnement de l'utilisateur connecté
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/Coach"
                    - type: object
                      properties:
                        subscribed: { type: boolean }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    Error:
      description: Erreur
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string, example: TEMPLATE_NOT_FOUND }
    TypeCircuit:
      type: string
      enum: [FORCE, ENDURANCE, HYPERTROPHIE, FIGURE]
    Profil:
      type: object
      properties:
        id: { type: integer }
        nom: { type: string }
        prenom: { type: string }
        mail: { type: string }
        poids: { type: number, nullable: true }
        taille: { type: number, nullable: true }
        age: { type: integer, nullable: true }
        blessures: { type: string, nullable: true }
        role: { type: string, enum: [USER, COACH, ADMIN] }
        createdAt: { type: string, format: date-time }
    ProfilInput:
      type: object
      properties:
        nom: { type: string }
        prenom: { type: string }
        mail: { type: string }
        motDePasse: { type: string }
        poids: { type: number }
        taille: { type: number }
        age: { type: integer }
        blessures: { type: string }
        role: { type: string, enum: [USER, COACH, ADMIN] }
    Coach:
      type: object
      properties:
        id: { type: integer }
        nom: { type: string }
        prenom: { type: string }
        role: { type: string, enum: [COACH, ADMIN] }
        createdAt: { type: string, format: date-time }
    ExerciceInput:
      type: object
      required: [exercice]
      properties:
        exercice: { type: string, description: "Nom du catalogue ou nom personnalisé", example: Squat }
        description: { type: string, nullable: true }
        consignes: { type: string, nullable: true }
        objectif: { type: string, nullable: true }
        series: { type: integer, minimum: 0, nullable: true }
        nbRep: { type: integer, minimum: 0, nullable: true }
        duree: { type: integer, minimum: 0, description: Secondes, nullable: true }
        tempsDeRepos: { type: integer, minimum: 0, description: Secondes, nullable: true }
        tempo: { type: string, nullable: true }
        lest: { type: number, minimum: 0, nullable: true }
        amplitude: { type: string, nullable: true }
        url: { type: string, nullable: true }
        ordre: { type: integer }
    Exercice:
      allOf:
        - $ref: "#/components/schemas/ExerciceInput"
        - type: object
          properties:
            id: { type: integer }
            circuitId: { type: integer }
    CircuitInput:
      type: object
      required: [nom, type, date]
      properties:
        nom: { type: string }
        type: { $ref: "#/components/schemas/TypeCircuit" }
        niveau: { type: string, nullable: true, description: "Difficulté en texte libre (ex. Débutant, ou une valeur personnalisée)" }
        date: { type: string, format: date }
    Circuit:
      allOf:
        - $ref: "#/components/schemas/CircuitInput"
        - type: object
          properties:
            id: { type: integer }
            auteurId: { type: integer }
            exercices: { type: array, items: { $ref: "#/components/schemas/Exercice" } }
            personnes:
              type: array
              items:
                type: object
                properties:
                  profilId: { type: integer }
    TemplateInput:
      type: object
      required: [nom, exercices]
      properties:
        nom: { type: string, example: Full body débutant }
        niveau: { type: string, nullable: true, description: "Difficulté en texte libre" }
        type: { allOf: [{ $ref: "#/components/schemas/TypeCircuit" }], nullable: true }
        exercices:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ExerciceInput" }
    TemplateSummary:
      type: object
      properties:
        id: { type: integer }
        nom: { type: string }
        niveau: { type: string, nullable: true }
        type: { allOf: [{ $ref: "#/components/schemas/TypeCircuit" }], nullable: true }
        coachId: { type: integer }
        exercicesCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Template:
      type: object
      properties:
        id: { type: integer }
        nom: { type: string }
        niveau: { type: string, nullable: true }
        type: { allOf: [{ $ref: "#/components/schemas/TypeCircuit" }], nullable: true }
        coachId: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        exercices:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/ExerciceInput"
              - type: object
                properties:
                  id: { type: integer }
                  templateId: { type: integer }